move the canonical StorageController implementation into mz_dataflow
