+++ src/materialized/Cargo.toml
@@ -56,6 +56,7 @@ libc = "0.2.122"
 mz-build-info = { path = "../build-info" }
 mz-coord = { path = "../coord" }
 mz-dataflow = { path = "../dataflow" }
+mz-storage = { path = "../storage" }
 mz-dataflow-types = { path = "../dataflow-types" }
 mz-frontegg-auth = { path = "../frontegg-auth" }
 mz-http-proxy = { path = "../http-proxy", features = ["reqwest", "hyper"] }
